+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes
 - Add --debug option to magik-lint
 - Fix ScopeBuilderVisitor for \_try without an identifier
 - Refactorings in MagikLint
+- Add check DuplicateMethodInFile
 
 
 0.3.2 (2019-10-29)